requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.22.0,<0.23
mysqlclient>=2.1.0,<3.0    # MySQL DB driver for Django
